// src/types/racing.ts
export type CategoryID =
  | '9daef0d7-bf3c-4f50-921d-8e818c60fe61' // Greyhound
  | '161d9be2-e909-4326-8c2c-35ed71fb460b' // Harness
  | '4a2788f8-e825-4d36-9894-efd4baf1cfae' // Horse

export const CATEGORY_LABELS: Record<CategoryID, string> = {
  '9daef0d7-bf3c-4f50-921d-8e818c60fe61': 'Greyhound',
  '161d9be2-e909-4326-8c2c-35ed71fb460b': 'Harness',
  '4a2788f8-e825-4d36-9894-efd4baf1cfae': 'Horse',
}

export type Race = {
  id: string
  meetingName: string
  raceNumber: number
  startMs: number
  categoryId: CategoryID
}